+++ lib/src/widgets/default_styles.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:tuple/tuple.dart';
 
 import '../../flutter_quill.dart';
-import '../../models/documents/style.dart';
+import '../models/documents/style.dart';
+import '../utils/platform.dart';
 
 class QuillStyles extends InheritedWidget {
   const QuillStyles({
@@ -192,19 +193,10 @@ class DefaultStyles {
     );
     const baseSpacing = Tuple2<double, double>(6, 0);
     String fontFamily;
-    switch (themeData.platform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        fontFamily = 'Menlo';
-        break;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.windows:
-      case TargetPlatform.linux:
-        fontFamily = 'Roboto Mono';
-        break;
-      default:
-        throw UnimplementedError();
+    if (isAppleOS(themeData.platform)) {
+      fontFamily = 'Menlo';
+    } else {
+      fontFamily = 'Roboto Mono';
     }
 
     final inlineCodeStyle = TextStyle(