+++ lib/src/widgets/raw_editor/raw_editor_state_selection_delegate_mixin.dart
@@ -3,7 +3,8 @@ import 'dart:math' as math;
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
-import '../../../utils/diff_delta.dart';
+import '../../models/documents/nodes/leaf.dart';
+import '../../utils/delta.dart';
 import '../editor.dart';
 
 mixin RawEditorStateSelectionDelegateMixin on EditorState
@@ -23,19 +24,33 @@ mixin RawEditorStateSelectionDelegateMixin on EditorState
 
     widget.controller.replaceText(
         diff.start, diff.deleted.length, insertedText, value.selection);
+
+    if (insertedText == pastePlainText && pastePlainText != '') {
+      final pos = diff.start;
+      for (var i = 0; i < pasteStyle.length; i++) {
+        final offset = pasteStyle[i].item1;
+        final style = pasteStyle[i].item2;
+        widget.controller.formatTextStyle(
+            pos + offset,
+            i == pasteStyle.length - 1
+                ? pastePlainText.length - offset
+                : pasteStyle[i + 1].item1,
+            style);
+      }
+    }
   }
 
   String _adjustInsertedText(String text) {
-    // For clip from editor, it may contain image, a.k.a 65532.
+    // For clip from editor, it may contain image, a.k.a 65532 or '\uFFFC'.
     // For clip from browser, image is directly ignore.
     // Here we skip image when pasting.
-    if (!text.codeUnits.contains(65532)) {
+    if (!text.codeUnits.contains(Embed.kObjectReplacementInt)) {
       return text;
     }
 
     final sb = StringBuffer();
     for (var i = 0; i < text.length; i++) {
-      if (text.codeUnitAt(i) == 65532) {
+      if (text.codeUnitAt(i) == Embed.kObjectReplacementInt) {
         continue;
       }
       sb.write(text[i]);